Resolve error on Linux when downloading packages
As of the latest release (1.2.1.0), the `css` command will throw the error, 

"Error: Specified file could not be compiled.

Could not find a part of the path '/home/user/.nuget/packages'."

This is because the directory "/home/user/" does not exist. Rather, ".nuget/packages" exists within "~/" as a hidden folder.
As such, "/home/user/.nuget/packages" should be changed to "~/.nuget/packages" to resolve the error.

This resolves https://github.com/oleg-shilo/cs-script/issues/195.